some more updates, should have all the testing proper now

+++ resources/catalog.rb
@@ -1,33 +1,48 @@
-property :catalog_config, Hash, required: true
+
+property :catalog_config, Hash, default: {
+  dbname: 'bareos',
+  dbuser: 'bareos',
+  dbpassword: '',
+}
 property :catalog_backend, String, equal_to: %w(postgresql mysql), default: 'postgresql'
 property :template_name, String, default: 'catalog.erb'
 property :template_cookbook, String, default: 'bareos'
 
+default_action :create
+
 action :create do
-  package 'bareos-database-common'
-  package 'bareos-database-postgresql' if new_resource.catalog_backend == 'postgresql'
-  package 'bareos-database-mysql' if new_resource.catalog_backend == 'mysql'
+  # I am only inlcuding the bareos-director package here because it is a
+  # dependency of "bareos-dbcheck" which is called from the database scripts.
+  # This ideally shouldn't be the case but..it is.. So just in case, installing.
+  package %w(
+    bareos-director
+    bareos-database-common
+    bareos-database-tools
+  )
+
+  case new_resource.catalog_backend
+  when 'mysql'
+    package 'bareos-database-mysql'
+    script_prefix = ''
+  else
+    package 'bareos-database-postgresql'
+    script_prefix = 'su postgres -s /bin/bash -c '
+  end
 
   template "director_catalog_#{new_resource.name}" do
-    source new_resource.template_name
     path "/etc/bareos/bareos-dir.d/catalog/#{new_resource.name}.conf"
+    source new_resource.template_name
     cookbook new_resource.template_cookbook
     owner 'bareos'
     group 'bareos'
     mode '0640'
     variables(
-      catalog_config: new_resource.catalog_config,
       catalog_name: new_resource.name,
-      catalog_backend: new_resource.catalog_backend
+      catalog_backend: new_resource.catalog_backend,
+      catalog_config: new_resource.catalog_config
     )
     action :create
   end
-
-  script_prefix = if new_resource.catalog_backend == 'postgresql'
-                    'su postgres -s /bin/bash -c '
-                  else
-                    ''
-                  end
 
   group 'bareos' do
     action :modify
@@ -36,38 +51,60 @@ action :create do
     only_if { new_resource.catalog_backend == 'postgresql' }
   end
 
-  execute "create_bareos_database_#{new_resource.name}" do
-    command <<-DBCREATED
-#{script_prefix}/usr/lib/bareos/scripts/create_bareos_database && \
-touch /var/lib/bareos/.dbcreated_#{new_resource.name}
+  bash "create_bareos_database_#{new_resource.name}" do
+    code <<-DBCREATED
+#!/bin/bash
+set -e
+#{script_prefix}/usr/lib/bareos/scripts/create_bareos_database
+RESULT=$?
+if [ $RESULT -eq 0 ]; then
+  su bareos -s /bin/bash -c "touch /var/lib/bareos/.dbcreated_#{new_resource.name}"
+fi
 DBCREATED
     creates "/var/lib/bareos/.dbcreated_#{new_resource.name}"
     action :run
   end
 
-  execute "make_bareos_tables_#{new_resource.name}" do
-    command <<-TABCREATED
-#{script_prefix}/usr/lib/bareos/scripts/make_bareos_tables && \
-touch /var/lib/bareos/.dbtabcreated_#{new_resource.name}
+  bash "make_bareos_tables_#{new_resource.name}" do
+    code <<-TABCREATED
+#!/bin/bash
+set -e
+#{script_prefix}/usr/lib/bareos/scripts/make_bareos_tables
+RESULT=$?
+if [ $RESULT -eq 0 ]; then
+  su bareos -s /bin/bash -c "touch /var/lib/bareos/.dbtabcreated_#{new_resource.name}"
+fi
 TABCREATED
     creates "/var/lib/bareos/.dbtabcreated_#{new_resource.name}"
     action :run
   end
 
-  execute "update_bareos_tables_#{new_resource.name}" do
-    command <<-TABUPDATED
-rm -f /var/lib/bareos/.dbprivgranted_#{new_resource.name}
-#{script_prefix}/usr/lib/bareos/scripts/update_bareos_tables && \
-touch /var/lib/bareos/.dbtabupdated_#{new_resource.name}
+  bash "update_bareos_tables_#{new_resource.name}" do
+    code <<-TABUPDATED
+#!/bin/bash
+set -e
+if [ -f /var/lib/bareos/.dbprivgranted_#{new_resource.name} ]; then
+  rm -f /var/lib/bareos/.dbprivgranted_#{new_resource.name}
+fi
+#{script_prefix}/usr/lib/bareos/scripts/update_bareos_tables
+RESULT=$?
+if [ $RESULT -eq 0 ]; then
+  su bareos -s /bin/bash -c "touch /var/lib/bareos/.dbtabupdated_#{new_resource.name}"
+fi
 TABUPDATED
     creates "/var/lib/bareos/.dbtabupdated_#{new_resource.name}"
     action :run
   end
 
-  execute "grant_bareos_privileges_#{new_resource.name}" do
-    command <<-PRIVGRANTED
-#{script_prefix}/usr/lib/bareos/scripts/grant_bareos_privileges && \
-touch /var/lib/bareos/.dbprivgranted_#{new_resource.name}
+  bash "grant_bareos_privileges_#{new_resource.name}" do
+    code <<-PRIVGRANTED
+#!/bin/bash
+set -e
+#{script_prefix}/usr/lib/bareos/scripts/grant_bareos_privileges
+RESULT=$?
+if [ $RESULT -eq 0 ]; then
+  su bareos -s /bin/bash -c "touch /var/lib/bareos/.dbprivgranted_#{new_resource.name}"
+fi
 PRIVGRANTED
     creates "/var/lib/bareos/.dbprivgranted_#{new_resource.name}"
     action :run
